+++ src/utility/detail/Compressor/NoneCompressor.cpp
@@ -10,12 +10,12 @@ NoneCompressor::~NoneCompressor() {}
 
 int64_t NoneCompressor::decompress(const UByteArrayAdapter& src, UByteArrayAdapter & dest) const {
 	dest.putData(0, src);
-	return src.size();
+	return (int64_t) src.size();
 }
 
 int64_t NoneCompressor::compress(const UByteArrayAdapter& src, UByteArrayAdapter & dest) const {
 	dest.putData(0, src);
-	return src.size();
+	return (int64_t) src.size();
 }
 
 }}}//end namespace sserialize::detail::Compressor
